pyproject.toml
[tool.poetry]
name = "quizybot"
version = "0.1.0"
description = ""
authors = ["Allan Delautre <allandela@hotmail.fr>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
bs4 = "^0.0.2"
thefuzz = "^0.22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
